More cleanup of code
