+++ API/Services/LocalizationService.cs
@@ -4,20 +4,14 @@ using System.Linq;
 using System.Text.Json;
 using System.Threading.Tasks;
 using API.Data;
+using API.DTOs;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Hosting;
 
 namespace API.Services;
 #nullable enable
 
-public class KavitaLocale
-{
-    public string FileName { get; set; } // Key
-    public string RenderName { get; set; }
-    public float TranslationCompletion { get; set; }
-    public bool IsRtL { get; set; }
-    public string Hash { get; set; } // ETAG hash so I can run my own localization busting implementation
-}
+
 
 
 public interface ILocalizationService
